+++ test/smokes/code_sniffer/expectations.rb
@@ -105,3 +105,52 @@ s.add_test(
   ],
   analyzer: { name: "PHP_CodeSniffer", version: default_version }
 )
+
+s.add_test(
+  "option_extensions",
+  type: "success",
+  issues: [
+    {
+      path: "foo.fcgi",
+      location: { start_line: 2, start_column: 1 },
+      id: "PSR2.Files.ClosingTag.NotAllowed",
+      message: "A closing tag is not permitted at the end of a PHP file",
+      links: [],
+      object: { type: "ERROR", severity: 5, fixable: true },
+      git_blame_info: {
+        commit: :_, line_hash: "48ee9fdb6490aadc28bbb90cee75350be54532de", original_line: 2, final_line: 2
+      }
+    },
+    {
+      path: "foo.inc",
+      location: { start_line: 2, start_column: 1 },
+      id: "PSR2.Files.ClosingTag.NotAllowed",
+      message: "A closing tag is not permitted at the end of a PHP file",
+      links: [],
+      object: { type: "ERROR", severity: 5, fixable: true },
+      git_blame_info: {
+        commit: :_, line_hash: "48ee9fdb6490aadc28bbb90cee75350be54532de", original_line: 2, final_line: 2
+      }
+    }
+  ],
+  analyzer: { name: "PHP_CodeSniffer", version: default_version }
+)
+
+s.add_test(
+  "option_ignore",
+  type: "success",
+  issues: [
+    {
+      path: "src/foo.php",
+      location: { start_line: 2, start_column: 1 },
+      id: "PSR2.Files.ClosingTag.NotAllowed",
+      message: "A closing tag is not permitted at the end of a PHP file",
+      links: [],
+      object: { type: "ERROR", severity: 5, fixable: true },
+      git_blame_info: {
+        commit: :_, line_hash: "48ee9fdb6490aadc28bbb90cee75350be54532de", original_line: 2, final_line: 2
+      }
+    }
+  ],
+  analyzer: { name: "PHP_CodeSniffer", version: default_version }
+)